Fix compilation warning
Decimal literal assigned to bytesXX variable will be left-aligned.
"bytesXX types are left-aligned (i.e. right-padded with zeros), which is intuitive for things like bytes32 a = "abcdef";, but it is less intuitive for bytes32 a = 7;. After all, bytesXX types are not really meant to be used for decimal numbers anyway. We should issue a warning whenever a decimal literal is used in a bytesXX context."
- https://github.com/ethereum/solidity/issues/2349